+++ docs_src/src/guide/changelog.md
@@ -16,6 +16,7 @@ meta:
 
 - Type-dependent rules validating integers (via the "integer" type rule) now share their validation message with the "numeric" type.
 - Added [paginators](./basics/database.html#pagination).
+- Added [`helper.EscapeLike()`](./advanced/helpers.html#helper-escapelike).
 
 ## v3.3.x